🚧 Move query parameter handling to Header
