+++ src/components/Header.js
@@ -1,33 +1,44 @@
-import React from "react";
+import React, { Component } from "react";
+import { withRouter } from "react-router-dom";
 import PropTypes from "prop-types";
 import Searchbar from "./Searchbar";
 import "../css/Header.scss";
 
-function Header({
-  username, onSignOutClick, toggleSidebar, setSearchbarValue, searchbarValue, searchHandler,
-}) {
-  return (
-    <header id="app-header">
-      <button id="toggle-btn" onClick={toggleSidebar}>
-        <i className="fa fa-bars" />
-      </button>
-      <Searchbar
-        value={searchbarValue}
-        searchHandler={searchHandler}
-        setSearchbarValue={setSearchbarValue}
-      />
-      <div id="user-info">
-        <div className="user-img" />
-        {username}
-        <button onClick={onSignOutClick}>
-          <i className="fa fa-sign-out-alt" /> Sign out
+class Header extends Component {
+  componentDidMount() {
+    const url = new URLSearchParams(this.props.location.search);
+    const query = url.get("query");
+    this.props.setSearchbarValue(query);
+  }
+
+  render() {
+    const {
+      username, onSignOutClick, toggleSidebar, setSearchbarValue, searchbarValue, searchHandler,
+    } = this.props;
+    return (
+      <header id="app-header">
+        <button id="toggle-btn" onClick={toggleSidebar}>
+          <i className="fa fa-bars" />
         </button>
-      </div>
-    </header>
-  );
+        <Searchbar
+          value={searchbarValue}
+          search={searchHandler}
+          setSearchbarValue={setSearchbarValue}
+        />
+        <div id="user-info">
+          <div className="user-img" />
+          {username}
+          <button onClick={onSignOutClick}>
+            <i className="fa fa-sign-out-alt" /> Sign out
+          </button>
+        </div>
+      </header>
+    );
+  }
 }
 
 Header.propTypes = {
+  location: PropTypes.object.isRequired, // from react-router
   username: PropTypes.string.isRequired,
   toggleSidebar: PropTypes.func.isRequired,
   onSignOutClick: PropTypes.func.isRequired,
@@ -36,4 +47,4 @@ Header.propTypes = {
   setSearchbarValue: PropTypes.func.isRequired,
 };
 
-export default Header;
+export default withRouter(Header);
